pyproject.toml
[project]
name = "fastapi-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.3.2",
    "alembic>=1.18.4",
    "asgiref>=3.11.1",
    "asyncpg>=0.31.0",
    "bcrypt>=4.0.0,<5.0.0",
    "celery>=5.6.3",
    "fastapi-mail>=1.6.2",
    "fastapi[standard]>=0.128.3",
    "flower>=2.0.1",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.12.1",
    "pytest>=9.0.2",
    "redis>=7.4.0",
    "schemathesis>=4.14.3",
    "sqlalchemy[asyncio]>=2.0.46",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = ["rich>=14.3.2"]
